fix start script
