+++ src/app.ts
@@ -1,0 +1,28 @@
+import cors from "cors";
+import express, { Request, Response } from "express";
+
+const app: express.Application = express();
+app.use(express.json());
+app.use(cors());
+app.use(express.urlencoded({ extended: true }));
+
+app.get("/", async (req: Request, res: Response): Promise<Response> => {
+  return res.status(200).send({
+    message: "Hello World!",
+  });
+});
+
+// test response - should be deleted.
+app.get("/name", async (req: Request, res: Response): Promise<Response> => {
+  return res.send("hi");
+});
+
+// try {
+//   app.listen(PORT, (): void => {
+//     console.log(`Connected successfully on port ${PORT}`);
+//   });
+// } catch (error: any) {
+//   console.error(`Error : ${error.message}`);
+// }
+
+export default app;
